etl/requirements.txt
async-timeout==4.0.2
certifi==2023.5.7
elastic-transport==8.4.0
elasticsearch==8.8.0
loguru==0.7.0
psycopg2-binary==2.9.6
pydantic==1.10.9
python-dotenv==1.0.0
redis==4.5.5
typing_extensions==4.6.3
urllib3==1.26.16
